smart-waste-backend-v2/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
slowapi
Pillow
bcrypt
pydantic[email]
ortools
qrcode[pil]
websockets
aiofiles
psutil
numpy
